Finished all 4 levels of decompression. Added selection of outfile name to gui. Other minor changes.

+++ src/main/java/com/example/skdecomp/SkFile.java
@@ -39,7 +39,7 @@ public class SkFile extends File {
         return checksum;
     }
 
-    public long getDictLength(){
+    public int getDictLength(){
         return dictLength;
     }
